+++ e2e-tests/tests/react-native.test.ts
@@ -11,8 +11,7 @@ import { afterAll, beforeAll, describe, test, expect } from 'vitest';
 //@ts-expect-error - clifty is ESM only
 import { KEYS, withEnv } from 'clifty';
 
-// TODO: remove skipIf when macOS CI is fixed
-describe.skipIf(process.platform === 'darwin')('ReactNative', () => {
+describe('ReactNative', () => {
   const integration = Integration.reactNative;
   let wizardExitCode: number;
   const { projectDir, cleanup } = createIsolatedTestEnv(
@@ -145,6 +144,10 @@ defaults.project=${TEST_ARGS.PROJECT_SLUG}
 
 defaults.url=https://sentry.io/`,
     );
+  });
+
+  test('ios/Podfile is updated with use_modular_headers!', () => {
+    checkFileContents(`${projectDir}/ios/Podfile`, `use_modular_headers!`);
   });
 
   test('build.gradle is updated correctly', () => {